extract low level utils from Spotify Client
