+++ src/server/services/spotify.ts
@@ -1,6 +1,7 @@
 import {env} from "~/env";
 import {redirect} from "next/navigation";
 import {stringify} from "node:querystring";
+import {addTimeInterval, generateRandomAlphanumString} from "~/lib/utils";
 
 
 enum SpotifyResponse {
@@ -46,7 +47,7 @@ class SpotifyClient {
     }
 
     initiateUserAuth() {
-        this.state = this.generateRandomString();
+        this.state = generateRandomAlphanumString();
         const authRequestBody = {
             response_type: SpotifyResponse.CODE,
             client_id: this.clientId,
@@ -69,9 +70,9 @@ class SpotifyClient {
                     method: 'POST',
                     headers: {
                         'content-type': 'application/x-www-form-urlencoded',
-                        Authorization: `Basic ${Buffer.from(
-                            this.clientId + ':' + this.clientSecret
-                        ).toString('base64')}`,
+                        Authorization:
+                            `Basic ${Buffer.from(this.clientId + ':' + this.clientSecret)
+                            .toString('base64')}`,
                     },
                     body: new URLSearchParams({
                         code: authorizationCode,
@@ -90,49 +91,6 @@ class SpotifyClient {
         }
     }
 
-
-    private generateRandomString (length = 64): string {
-        let text = '';
-        const possible =
-            'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-
-        for (let i = 0; i < length; i++) {
-            text += possible.charAt(Math.floor(Math.random() * possible.length));
-        }
-        return encodeURIComponent(text);
-    };
-
-}
-
-type TimeInterval = 'years' | 'months' | 'days' | 'hours' | 'minutes' | 'seconds';
-
-function addTimeInterval(date: Date, amount: number, interval: TimeInterval): Date {
-    const newDate = new Date(date.getTime()); // Create a new date object to avoid mutating the original date
-
-    switch (interval) {
-        case 'years':
-            newDate.setFullYear(date.getFullYear() + amount);
-            break;
-        case 'months':
-            newDate.setMonth(date.getMonth() + amount);
-            break;
-        case 'days':
-            newDate.setDate(date.getDate() + amount);
-            break;
-        case 'hours':
-            newDate.setHours(date.getHours() + amount);
-            break;
-        case 'minutes':
-            newDate.setMinutes(date.getMinutes() + amount);
-            break;
-        case 'seconds':
-            newDate.setSeconds(date.getSeconds() + amount);
-            break;
-        default:
-            throw new Error(`Invalid interval.`);
-    }
-
-    return newDate;
 }
